requirements.txt
streamlit
genai_agent_service_bmc_python_client-0.1.77-py2.py3-none-any.whl
oci
langchain-community
tiktoken